Possible fix for buying square bug
Possible (untested) fix for when two players land on the same square and the second player buys it. If a player (owner) buys a square that another player is already on, but who had declined or couldn’t buy it, they would still pay the owner before they rolled on their turn, even though they landed on the square first. Players should only have to pay the owner if they land on the square after it is bought.